requirements.txt
# Trademark Monitor Requirements
# ==============================

# Core dependencies
requests>=2.28.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Database
# (SQLite is built into Python)

# Dashboard
streamlit>=1.29.0
pandas>=2.0.0

# For scheduled tasks
schedule>=1.2.0

# Optional: For better XML processing
lxml>=4.9.0

# Optional: For testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Optional: For development
black>=23.0.0
flake8>=6.0.0
